pyproject.toml
[project]
name = "chainsight"
version = "0.1.0"
description = "Multi-agent supply chain disruption prediction system using Google ADK + A2A"
requires-python = ">=3.12"
dependencies = [
    "google-adk[a2a]>=1.30.0",
    "google-genai>=1.0.0",
    "uvicorn>=0.30.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "pydantic>=2.0.0",
    "firebase-admin>=6.0.0",
    "feedparser>=6.0.0",
    "networkx>=3.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
